Add windows build through appveyor (#9)
Adding appveyor.yml so we can do CI in a windows environment.

The change also replaces the rust-cypto crate with the md5
crate for doing md5 hashing within this crate. This allows this
crate to be more portable because the rust-crypto crate
requires additional work to get it working on windows.